requirements.txt
# CommerceBox - Sistema de Inventario Comercial Dual
# Dependencias del proyecto - VERSIÓN CORREGIDA

# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-extensions==3.2.3

# Base de datos
psycopg2-binary==2.9.9
redis==4.6.0

# Tareas asíncronas
celery==5.3.4
celery[redis]==5.3.4
django-celery-beat==2.5.0

# Autenticación y seguridad
django-allauth==0.57.0
djangorestframework-simplejwt==5.3.0
cryptography==41.0.7
bcrypt==4.1.2

# APIs y servicios web
requests==2.31.0
httpx==0.25.2

# Códigos de barras y reportes
python-barcode==0.15.1
Pillow==10.1.0
reportlab==4.0.7
openpyxl==3.1.2

# Hardware integration
pyserial==3.5
python-escpos==3.0

# Utilidades
python-decouple==3.8
python-dateutil==2.8.2
pytz==2023.3

# Filtros y paginación
django-filter==23.5

# Validadores adicionales
django-phonenumber-field==7.3.0
phonenumbers==8.13.26

# Email y notificaciones
django-anymail==10.2

# Desarrollo y testing
pytest==7.4.3
pytest-django==4.6.0
factory-boy==3.3.0
coverage==7.3.2
pytest-cov==4.1.0
requests-mock==1.11.0

# Formato y calidad de código
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoreo y logs
sentry-sdk==1.38.0
django-debug-toolbar==4.2.0
structlog==23.2.0
django-structlog==7.0.0

# Facturación electrónica
lxml==4.9.3
xmltodict==0.13.0

# Optimización de imágenes
django-imagekit==5.0.0

# Internacionalización
django-rosetta==0.9.9

# Backup y exportación
django-dbbackup==4.0.2

# Rate limiting y caching
django-ratelimit==4.1.0
django-redis==5.4.0

# Seguridad adicional
# django-security==0.21.0  # ELIMINADO: Esta versión no existe, máxima disponible es 0.12.0
django-permissions-policy==4.17.0
django-csp==3.7

# Timezone management
django-timezone-field==6.1.0

# Environment and configuration
django-environ==0.11.2

# Session management
django-user-sessions==2.0.0

# API documentation
drf-spectacular==0.27.0

# Health checks
django-health-check==3.17.0

# Validación de datos adicional
cerberus==1.3.5
marshmallow==3.20.2

# Generación de UUIDs
shortuuid==1.0.11

# Gestión de archivos temporales
django-cleanup==8.0.0

# Compresión de respuestas
django-compression-middleware==0.5.0

# Middleware de seguridad adicional
django-axes==6.1.1

# Exportación y reportes mejorados
tablib==3.5.0
xlwt==1.3.0

# WebSocket support para notificaciones en tiempo real
channels==4.0.0
channels-redis==4.1.0

# Monitoring y métricas
django-prometheus==2.3.1
prometheus-client==0.19.0
django-filter==23.5
 django-redis==5.4.0